Migrate Subscribe Preview Page to Inertia (#3167)
Issue: https://github.com/antiwork/gumroad/issues/3149
Closes: https://github.com/antiwork/gumroad/issues/3149

- [x]  Migrate `GET /:username/subscribe_preview` to render by Inertia
- [x] Ensure generate_subscribe_preview works

As per the initiatiative of migrating all pages to Inertia, We want to
migrate SubbscribePreview to it as well.

Use inertia instead of react_on_rails

---

[Before.webm](https://github.com/user-attachments/assets/45158312-e11a-4d78-bf0a-e2bdcb65a1dc)

[After.webm](https://github.com/user-attachments/assets/92694e4b-6946-443d-93dd-67983e56d43d)

---

<img width="1918" height="801" alt="controller tests"
src="https://github.com/user-attachments/assets/cc78bacf-e5e9-42d2-b39c-3ab6da2c1411"
/>

<img width="1918" height="667" alt="image"
src="https://github.com/user-attachments/assets/632fb66e-d51a-43fa-a104-f0a83fa6a9ae"
/>

<img width="1024" height="258" alt="before"
src="https://github.com/user-attachments/assets/f730fdd9-9992-4522-bd40-2823c97c5422"
/>

<img width="1024" height="258" alt="after"
src="https://github.com/user-attachments/assets/b8ba6bf9-8faf-4301-94fe-66e2cda55f42"
/>

---

- [x] I have read the [contributing
guidelines](https://github.com/antiwork/gumroad/blob/main/CONTRIBUTING.md)
- [x] I have watched [Gumroad PR review
livestreams](https://www.youtube.com/@anti-work)
- [x] I have performed a self-review and left review comments on my PR
- [x] I have added/updated tests for my changes

---

Cursor for auto-completion only.

---------
