--- app/views/users/subscribe_preview.html.erb
@@ -1,2 +1,0 @@
-<%= render("layouts/custom_styles/style") %>
-<%= react_component "SubscribePreview", props: @subscribe_preview_props, prerender: true %>
